Scan through a set of flycircuit neurons

Can also choose to select specific neurons along the way and navigate forwards and backwards. NB this is simply a wrapper for nat::nlscan, with the additional function of converting all neuron identifiers to standard flycircuit identifiers.

dpscan(neurons, db = NULL, col = "red", Verbose = T, Wait = T,
  sleep = 0.1, extrafun = NULL, selected_file = NULL,
  selected_col = "green", yaml = TRUE, ...)

Arguments

neurons
vector of flycircuit identifiers to plot *(anything that fc_gene_name understands)

db
the neuronlist containing the neurons. Defaults to the list named by options("nat.default.neuronlist").

col
the colour with which to plot the neurons (default 'red').

Verbose
logical indicating that info about each selected neuron should be printed (default TRUE).

Wait
logical indicating that there should be a pause between each displayed neuron.

sleep
time to pause between each displayed neuron when Wait=TRUE.

extrafun
an optional function called when each neuron is plotted, with args gene_name and selected.

selected_file
an optional path to a yaml file that already contains a selection.

selected_col
the color in which selected neurons (such as those specified in selected_file) should be plotted.

yaml
a logical indicating that selections should be saved to disk in yaml rather than rda format.

...
extra arguments to pass to plot3dfc.

Value

A character vector of names of any selected neurons, of length 0 if none selected.

See also

nlscan, fc_gene_name

Examples

not_run({
  # numeric idids, specifying db explicitly
  dpscan(c(1024L, 10616L, 8399L), db=kcs20)

  # using an option to set db
  op<-options(nat.default.neuronlist='kcs20')
  dpscan(c("fru-M-500112", "Gad1-F-900005", "Gad1-F-100010"))
  options(op)

})