Antennal lobe glomerulus (where applicable) for FlyCircuit identifiers

These comes from manual annotations carried out for Costa et al.

fc_glom(x = NULL)

Arguments

x
A neuron identifier. When missing defaults to all neurons annotated with an ALGlomerulus tag.

Value

a character vector of glomerulus names, named by the neuron identifier

See also

fc_neuron_type, fc_gene_name

Examples

fc_glom('FruMARCM-F000446_seg001')

#> FruMARCM-F000446_seg001 
#>                  "VL2p" 

# how many neurons annotated with any glomerulus
length(fc_glom())

#> [1] 433

# how many neurons annotated for each glomerulus
table(fc_glom())

#> 
#>   D doublet         DA1     DA1+DL3         DA2         DA4         DC2 
#>           1          17           1           8           2           1 
#>         DC3         DL1        DL2d        DL2v   DL2v+DM3+         DL3 
#>           1          17         145          37           1           5 
#>         DL4         DM1    DM3+DP1l         DM5         DM6        DP1l 
#>           1           1           1           8          10           2 
#>       DP1l+       DP1m+           V          V+        VA1d    VA1d+VM5 
#>           1           1           1          14           3           1 
#>       VA1lm      VA1lm+         VA4        VA7m VAsomething         VC2 
#>          22           1           4           2           2           2 
#>     VC3+VM4        VC3l        VC3m         VC4         VL1        VL2a 
#>           1          11          10           4           1           2 
#>        VL2p         VM1         VM2         VM3         VM4        VM4+ 
#>           1          13           7           1           2           3 
#>    VM4+DM3+        VM5d        VM5v   VM5v+VC3l         VM6        VM6+ 
#>           2          20          13           1           1           1 
#>     VM6+VP2         VM7         VP?         VP1        VP2+     VP3+VP4 
#>           1          17           1           1           1           1 
#>         VP4 
#>           4