Return metadata including NeuronType for FlyCircuit neurons

fc_neuron_type depends on a more generic worker function fc_annotated_class, which may be used to construct other convenience functions.

fc_neuron_type(x = NULL, regex = NULL, ...)

fc_annotated_class(class, x = NULL, regex = NULL, ...)

Arguments

x
A neuron identifier. When missing defaults to all neurons annotated with the given annotation class.

regex
An optional regex used to filter the annotation text values of the matching neurons.

...
Additional arguments passed to grepl.

class
The annotation class to select (for codefc_annotated_class).

Value

a character vector of annotation values, named by the neuron identifier

Details

In our schema, annotations are key-value pairs in which an annotation class, such as NeuronType, may have multiple values (e.g. gamma Kenyon cell).

See also

grepl, fc_gene_name,

Examples

# how many neurons are annotated with any type?
length(fc_neuron_type())

#> [1] 7021

# how many types of neuron
length(unique(fc_neuron_type()))

#> [1] 200

# how many neurons of each type are present
table(fc_neuron_type())

#> 
#>                                                                 adPN 
#>                                                                  332 
#>                                            adult ALv1 lineage neuron 
#>                                                                    1 
#>                                            adult ALv2 lineage neuron 
#>                                                                    5 
#>                                          adult AOTUv2 lineage neuron 
#>                                                                   14 
#>                                          adult AOTUv4 lineage neuron 
#>                                                                    7 
#>                                             adult DM1 lineage neuron 
#>                                                                   30 
#>                                             adult DM3 lineage neuron 
#>                                                                   48 
#>                                             adult DM4 lineage neuron 
#>                                                                   50 
#>                                             adult DM6 lineage neuron 
#>                                                                  295 
#>                                         adult fruitless aDT-a neuron 
#>                                                                   35 
#>                                adult fruitless aDT-b (female) neuron 
#>                                                                   54 
#>                                  adult fruitless aDT-b (male) neuron 
#>                                                                   69 
#>                                adult fruitless aDT-e (female) neuron 
#>                                                                   21 
#>                                  adult fruitless aDT-e (male) neuron 
#>                                                                   45 
#>                                         adult fruitless aDT-f neuron 
#>                                                                   11 
#>                                         adult fruitless aDT-g neuron 
#>                                                                   10 
#>                                         adult fruitless aDT-h neuron 
#>                                                                   24 
#>                                         adult fruitless aIP-a neuron 
#>                                                                   24 
#>                                adult fruitless aIP-b (female) neuron 
#>                                                                    2 
#>                                  adult fruitless aIP-b (male) neuron 
#>                                                                    2 
#>                                adult fruitless aIP-c (female) neuron 
#>                                                                    2 
#>                                  adult fruitless aIP-c (male) neuron 
#>                                                                    3 
#>                                adult fruitless aIP-e (female) neuron 
#>                                                                   32 
#>                                  adult fruitless aIP-e (male) neuron 
#>                                                                   21 
#>                                adult fruitless aIP-g (female) neuron 
#>                                                                   12 
#>                                  adult fruitless aIP-g (male) neuron 
#>                                                                    2 
#>                                         adult fruitless aIP-h neuron 
#>                                                                   21 
#>                                         adult fruitless aSG-a neuron 
#>                                                                    2 
#>                                adult fruitless aSP-a (female) neuron 
#>                                                                    7 
#>                                  adult fruitless aSP-a (male) neuron 
#>                                                                   16 
#>                                adult fruitless aSP-b (female) neuron 
#>                                                                    4 
#>                                  adult fruitless aSP-b (male) neuron 
#>                                                                    7 
#>                                         adult fruitless aSP-d neuron 
#>                                                                   32 
#>                                adult fruitless aSP-g (female) neuron 
#>                                                                    6 
#>                                         adult fruitless aSP-j neuron 
#>                                                                    9 
#>                                adult fruitless aSP-k (female) neuron 
#>                                                                    9 
#>                                  adult fruitless aSP-k (male) neuron 
#>                                                                    1 
#>                                         adult fruitless aSP-l neuron 
#>                                                                    4 
#>                                         adult fruitless aSP-m neuron 
#>                                                                    6 
#>                                            adult fruitless Lo neuron 
#>                                                                   34 
#>                                             adult fruitless M neuron 
#>                                                                   98 
#>                                    adult fruitless Mb lineage neuron 
#>                                                                  682 
#>                                  adult fruitless pIP-a (male) neuron 
#>                                                                    3 
#>                                         adult fruitless pIP-b neuron 
#>                                                                   16 
#>                                         adult fruitless pIP-c neuron 
#>                                                                   69 
#>                                         adult fruitless pIP-d neuron 
#>                                                                   93 
#>                                adult fruitless pIP-e (female) neuron 
#>                                                                   22 
#>                                  adult fruitless pIP-e (male) neuron 
#>                                                                   15 
#>                                         adult fruitless pIP-g neuron 
#>                                                                    4 
#>                                adult fruitless pMP-a (female) neuron 
#>                                                                   40 
#>                                  adult fruitless pMP-a (male) neuron 
#>                                                                   47 
#>                                adult fruitless pMP-b (female) neuron 
#>                                                                   17 
#>                                  adult fruitless pMP-b (male) neuron 
#>                                                                   16 
#>                                adult fruitless pMP-c (female) neuron 
#>                                                                    2 
#>                                  adult fruitless pMP-c (male) neuron 
#>                                                                   33 
#>                                         adult fruitless pMP-d neuron 
#>                                                                    4 
#>                                adult fruitless pMP-e (female) neuron 
#>                                                                   29 
#>                                  adult fruitless pMP-e (male) neuron 
#>                                                                  207 
#>                                adult fruitless pMP-f (female) neuron 
#>                                                                   52 
#>                                  adult fruitless pMP-f (male) neuron 
#>                                                                   91 
#>                                         adult fruitless pSG-a neuron 
#>                                                                    8 
#>                                         adult fruitless pSP-a neuron 
#>                                                                   16 
#>                                adult fruitless pSP-b (female) neuron 
#>                                                                    2 
#>                                  adult fruitless pSP-b (male) neuron 
#>                                                                    7 
#>                                         adult fruitless pSP-c neuron 
#>                                                                    2 
#>                                            adult LHl3 lineage neuron 
#>                                                                   15 
#>                                            adult LHp1 lineage neuron 
#>                                                                    5 
#>                                          adult SLPpm2 lineage neuron 
#>                                                                    9 
#>                                         adult SMPp&v2 lineage neuron 
#>                                                                   25 
#>                                          adult SMPpd2 lineage neuron 
#>                                                                   41 
#>                                          adult SMPpv1 lineage neuron 
#>                                                                   21 
#>                                         adult VLPl&p2 lineage neuron 
#>                                                                   15 
#>                                         adult VLPp&l1 lineage neuron 
#>                                                                    2 
#>                                           adult VPNd1 lineage neuron 
#>                                                                    4 
#>                                           adult VPNd3 lineage neuron 
#>                                                                   10 
#>                                           adult VPNp2 lineage neuron 
#>                                                                   15 
#>                                           adult VPNv2 lineage neuron 
#>                                                                   35 
#>                                           adult WEDa2 lineage neuron 
#>                                                                    5 
#>                                                 AL Descending Neuron 
#>                                                                    1 
#>                          alpha'/beta' anterior-posterior Kenyon cell 
#>                                                                    1 
#>                                             alpha'/beta' Kenyon cell 
#>                                                                   85 
#>                                          alpha/beta core Kenyon cell 
#>                                                                  689 
#>                                               alpha/beta Kenyon cell 
#>                                                                    1 
#>                                     alpha/beta posterior Kenyon cell 
#>                                                                   35 
#>                                       alpha/beta surface Kenyon cell 
#>                                                                  366 
#>                                                              AMMC LN 
#>                                                                   33 
#>                                                        AMMC-AMMC PN1 
#>                                                                   34 
#>                                                        AMMC-AMMC PN2 
#>                                                                   21 
#>                                                           AMMC-B1 PN 
#>                                                                   99 
#>                                                        AMMC-IVLP PN1 
#>                                                                   78 
#>                                                        AMMC-IVLP PN2 
#>                                                                  170 
#>                                                          AMMC-VLP PN 
#>                                                                   15 
#>                                                          AMMC-VNC PN 
#>                                                                   12 
#>        antennal mechanosensory and motor center A1 projection neuron 
#>                                                                   16 
#>  antennal mechanosensory and motor center AMMC-AL1 projection neuron 
#>                                                                    1 
#> antennal mechanosensory and motor center AMMC-VLP1 projection neuron 
#>                                                                    3 
#> antennal mechanosensory and motor center AMMC-VLP2 projection neuron 
#>                                                                    1 
#> antennal mechanosensory and motor center AMMC-VLP3 projection neuron 
#>                                                                    1 
#> antennal mechanosensory and motor center AMMC-VLP5 projection neuron 
#>                                                                    1 
#>        antennal mechanosensory and motor center B2 projection neuron 
#>                                                                    1 
#>                                                BAmas2 lineage neuron 
#>                                                                    5 
#>                                                         bilateral PN 
#>                                                                    4 
#>                                                  BLD2 lineage neuron 
#>                                                                    1 
#>                                                  CCP-VMP-PB 2 neuron 
#>                                                                    1 
#>                                                  CCP-VMP-PB 4 neuron 
#>                                                                    1 
#>                                                          contra adPN 
#>                                                                    8 
#>                                                            contra PN 
#>                                                                    3 
#>                                                  ContraProjAcrossALs 
#>                                                                   12 
#>                                                      CSD interneuron 
#>                                                                    1 
#>                                              CVLP-IDFP-VMP-PB neuron 
#>                                                                    2 
#>                                                    descending neuron 
#>                                                                   17 
#>                          dFB neuron of the dopaminergic PPL1 cluster 
#>                                                                   15 
#>                  dorsal anterior lateral neuron of the protocerebrum 
#>                                                                    1 
#>                           DP neuron of the dopaminergic PPL1 cluster 
#>                                                                    5 
#>                           EB neuron of the dopaminergic PPM3 cluster 
#>                                                                   20 
#>                                    EB-IDFP DSB-PB 2 glomeruli neuron 
#>                                                                    1 
#>                                        EB-IDFP DSB-PB slice 3 neuron 
#>                                                                    1 
#>                                        EB-IDFP DSB-PB slice 7 neuron 
#>                                                                    1 
#>                             EB-IDFP VSB-PB slice 4 glomerulus neuron 
#>                                                                    1 
#>                             EB-IDFP VSB-PB slice 8 glomerulus neuron 
#>                                                                    1 
#>                                             gamma dorsal Kenyon cell 
#>                                                                   45 
#>                                                    gamma Kenyon cell 
#>                                                                  423 
#>                                               gamma main Kenyon cell 
#>                                                                    1 
#>          inferior ventrolateral protocerebrum local IVLP-IVLP neuron 
#>                                                                    1 
#>                                                              IVLP LN 
#>                                                                   49 
#>                                                         IVLP-IVLP PN 
#>                                                                    2 
#>                                         lobula columnar neuron Lcn10 
#>                                                                   84 
#>                                        lobula columnar neuron Lcn10A 
#>                                                                    8 
#>                                        lobula columnar neuron Lcn10B 
#>                                                                   86 
#>                                         lobula columnar neuron Lcn11 
#>                                                                   49 
#>                                         lobula columnar neuron Lcn12 
#>                                                                  212 
#>                                         lobula columnar neuron Lcn13 
#>                                                                    3 
#>                                         lobula columnar neuron Lcn14 
#>                                                                    7 
#>                                          lobula columnar neuron Lcn4 
#>                                                                   54 
#>                                          lobula columnar neuron Lcn6 
#>                                                                   30 
#>                                          lobula columnar neuron Lcn9 
#>                                                                   40 
#>                                        lobula tangential neuron Lt12 
#>                                                                   18 
#>                                                                  lPN 
#>                                                                  294 
#>                     MB-alpha neuron of the dopaminergic PPL1 cluster 
#>                                                                   51 
#>               MB-alpha prime neuron of the dopaminergic PPL1 cluster 
#>                                                                    6 
#>                       MB-AMP neuron of the dopaminergic PPL1 cluster 
#>                                                                    3 
#>                        MB-SV neuron of the dopaminergic PPL1 cluster 
#>                                                                   18 
#>                          mFB neuron of the dopaminergic PPM3 cluster 
#>                                                                   16 
#>               mushroom body medial-vertical lobe arborizing neuron 1 
#>                                                                    1 
#>             mushroom body pedunculus-medial lobe arborizing neuron 1 
#>                                                                    1 
#>                       mushroom body ventral lobe arborizing neuron 3 
#>                                                                   12 
#>                      mushroom body vertical lobe arborizing neuron 1 
#>                                                                    9 
#>                     PB 2 glomeruli-FB layer 2-IDFP HB-lateral neuron 
#>                                                                   33 
#>                PB 2 glomeruli-FB layers 1 to 4-IDFP HB-medial neuron 
#>                                                                    1 
#>                              PB slice 1-EB posterior-IDFP DSB neuron 
#>                                                                    1 
#>             PB slice 1-FB layer 2-IDFP HB-lateral ipsilateral neuron 
#>                                                                    3 
#>                         PB slice 1-FB layer 2-IDFP HB-lateral neuron 
#>                                                                   37 
#>          PB slice 1-FB layers 1 to 4-bilateral IDFP HB-medial neuron 
#>                                                                    2 
#>                     PB slice 1-FB layers 3 and 6-ventral gall neuron 
#>                                                                    2 
#>                       PB slice 1-FB layers layers 3-5-IDFP RB neuron 
#>                                                                    7 
#>                         PB slice 2-FB layer 2-IDFP HB-lateral neuron 
#>                                                                    1 
#>        PB slice 2-FB layers 1 to 4-IDFP HB-medial ipsilateral neuron 
#>                                                                    1 
#>                    PB slice 2-FB layers 1 to 4-IDFP HB-medial neuron 
#>                                                                    3 
#>                              PB slice 2-FB layers 3-5-IDFP RB neuron 
#>                                                                    3 
#>                   PB slice 2-FB-NO subunit 3 posterior domain neuron 
#>                                                                    2 
#>                         PB slice 3-FB layer 2-IDFP HB-lateral neuron 
#>                                                                    3 
#>                         PB slice 3-FB layer 5-IDFP HB-lateral neuron 
#>                                                                    1 
#>          PB slice 3-FB layers 1 to 4-bilateral IDFP HB-medial neuron 
#>                                                                   25 
#>                    PB slice 3-FB layers 1 to 4-IDFP HB-medial neuron 
#>                                                                    2 
#>                     PB slice 3-FB layers 3 and 6-ventral gall neuron 
#>                                                                    1 
#>                              PB slice 3-FB layers 3-5-IDFP RB neuron 
#>                                                                    3 
#>                    PB slice 3-FB-NO subunit 3 anterior domain neuron 
#>                                                                    2 
#>                   PB slice 3-FB-NO subunit 3 posterior domain neuron 
#>                                                                    1 
#>                    PB slice 4-FB layers 1 to 4-IDFP HB-medial neuron 
#>                                                                    1 
#>                     PB slice 4-FB layers 3 and 6-ventral gall neuron 
#>                                                                    1 
#>                              PB slice 4-FB layers 3-5-IDFP RB neuron 
#>                                                                   11 
#>                      PB slice 4-FB-NO subunit 2 dorsal domain neuron 
#>                                                                    1 
#>                   PB slice 4-FB-NO subunit 3 posterior domain neuron 
#>                                                                    6 
#>                         PB slice 5-FB layer 2-IDFP HB-lateral neuron 
#>                                                                    1 
#>                    PB slice 5-FB layers 1 to 4-IDFP HB-medial neuron 
#>                                                                    3 
#>                              PB slice 5-FB layers 3-5-IDFP RB neuron 
#>                                                                    1 
#>                   PB slice 5-FB-NO subunit 3 posterior domain neuron 
#>                                                                    4 
#>                         PB slice 6-FB layer 2-IDFP HB-lateral neuron 
#>                                                                   31 
#>                    PB slice 6-FB layers 1 to 4-IDFP HB-medial neuron 
#>                                                                    2 
#>                     PB slice 6-FB layers 3 and 6-ventral gall neuron 
#>                                                                    4 
#>                              PB slice 6-FB layers 3-5-IDFP RB neuron 
#>                                                                    4 
#>                      PB slice 6-FB-NO subunit 2 dorsal domain neuron 
#>                                                                    2 
#>                   PB slice 6-FB-NO subunit 3 posterior domain neuron 
#>                                                                    6 
#>                         PB slice 7-FB layer 2-IDFP HB-lateral neuron 
#>                                                                    1 
#>                    PB slice 7-FB layers 1 to 4-IDFP HB-medial neuron 
#>                                                                   41 
#>                     PB slice 7-FB layers 3 and 6-ventral gall neuron 
#>                                                                    6 
#>                              PB slice 7-FB layers 3-5-IDFP RB neuron 
#>                                                                    2 
#>                              PB slice 8-EB posterior-IDFP DSB neuron 
#>                                                                    1 
#>                     PB slice 8-FB layers 3 and 6-ventral gall neuron 
#>                                                                    3 
#>                    PB slice 8-FB-NO subunit 3 anterior domain neuron 
#>                                                                    1 
#>                   PB slice 8-FB-NO subunit 3 posterior domain neuron 
#>                                                                    6 
#>                   PB slice 9-FB-NO subunit 3 posterior domain neuron 
#>                                                                    6 
#>                                                    PB-EB-NO 3 neuron 
#>                                                                    2 
#>                                                    PB-EB-NO 5 neuron 
#>                                                                    2 
#>                      protocerebral bridge 18 glomeruli type 1 neuron 
#>                                                                   20 
#>                      protocerebral bridge 18 glomeruli type 2 neuron 
#>                                                                   49 
#>                        ventrolateral protocerebrum local VLP1 neuron 
#>                                                                   88 
#>                          vFB neuron of the dopaminergic PPM3 cluster 
#>                                                                   26 
#>                                                                 vlPN 
#>                                                                    2 
#>                                                                  vPN 
#>                                                                  179 

# find all the Kenyon cells
table(fc_neuron_type(regex="Kenyon"))

#> 
#> alpha'/beta' anterior-posterior Kenyon cell 
#>                                           1 
#>                    alpha'/beta' Kenyon cell 
#>                                          85 
#>                 alpha/beta core Kenyon cell 
#>                                         689 
#>                      alpha/beta Kenyon cell 
#>                                           1 
#>            alpha/beta posterior Kenyon cell 
#>                                          35 
#>              alpha/beta surface Kenyon cell 
#>                                         366 
#>                    gamma dorsal Kenyon cell 
#>                                          45 
#>                           gamma Kenyon cell 
#>                                         423 
#>                      gamma main Kenyon cell 
#>                                           1 

table(fc_neuron_type(regex="gamma Kenyon"))

#> 
#> gamma Kenyon cell 
#>               423 

table(fc_neuron_type(regex="gamma.*Kenyon"))

#> 
#> gamma dorsal Kenyon cell        gamma Kenyon cell   gamma main Kenyon cell 
#>                       45                      423                        1