fcidtable - Data.frame of flycircuit identifiers

The flycircuit database seems to have at least 3 unique identifiers for neurons. These are:

ididAn integer identifier which ranged from 1:16226 in the April 2011 release

gene_nameA character identifier of the general form "DvGlutMARCM-F003905_seg001"

NameA character identifier of the general form "VGlut-F-400794"

Details

Although the neuron Name is generally what is reported in publications the gene_name seems to widely used and is what connected to the confocal image file name. The _seg001 suffix in gene_names refers to the segmentation of individiual neurons from stacks that might contain multiple labelled neurons after MARCM.

See also

Other flycircuit-ids: fc_gene_name