Support functions for analysis of raw image data provided by flycircuit.tw

Support functions for analysis of raw image data provided by flycircuit.tw

Package options

flycircuit uses the following options to configure behaviour:

flycircuit.datadir: path to directory where downloaded data will be stored. Defaults to package's extdata folder.

flycircuit.dbdir: location to which dataframes caching database tables will be downloaded.

flycircuit.ffdir: location of ff objects used for all by all score matrices.

flycircuit.bigmatdir: location of big.matrix objects previously used for all by all score matrices.

flycircuit.scoremat: character vector naming the default regular or ff matrix object containing nblast scores. Note that although it is is permissible to set this to the actual score matrix rather than a character vector naming the matrix, this turns out to have substantial performance penalties for large matrices.

flycircuit.remote_overwrite: Determines how local cached versions of data available on the web are updated. When NULL (i.e. unset) the web version is checked and, if it is newer, it is donwloaded and used to overwrite the cached copy. When FALSE the exisiting local copy is used; when TRUE the remote version is always downloaded.

References

http://flycircuit.tw

See also

fc_download_data

Examples

# Show state of flycircuit package options
options()[grep('^flycircuit', names(options()))]

#> $flycircuit.bigmatdir
#> [1] "/Users/jefferis/Library/Application Support/rpkg-flycircuit/data/bigmatrix"
#> 
#> $flycircuit.datadir
#> [1] "/Users/jefferis/Library/Application Support/rpkg-flycircuit/data"
#> 
#> $flycircuit.dbdir
#> [1] "/Users/jefferis/Library/Application Support/rpkg-flycircuit/db"
#> 
#> $flycircuit.ffdir
#> [1] "/Users/jefferis/Library/Application Support/rpkg-flycircuit/data/ff"
#> 
#> $flycircuit.scoremat
#> [1] "kcs20scores"
#> 
#> $flycircuit.sidataurl
#> [1] "http://flybrain.mrc-lmb.cam.ac.uk/si/nblast/flycircuit"
#>