Convert between different FlyCircuit identifiers

fc_gene_name returns a character vector of gene_names of fly circuit neurons given idid or neuron Name fc_idid returns an integer idid of a fly circuit neuron given gene_name or neuron Name fc_neuron gets the neuron Name of a fly circuit neuron given gene_name or idid

fc_gene_name(x, ignore.case = TRUE)

fc_idid(x, ignore.case = TRUE)

fc_neuron(x, ignore.case = TRUE)

Arguments

x
Vector of flycircuit identifiers (integer idid, chaacter/factor gene_name/Name)

ignore.case
whether to ignore the case of the input.

Value

character vector of gene_name/Name or integer vector of idid.

Details

Depends on fcidtable dataframe

See also

Other flycircuit-ids: fcidtable

Examples

fc_gene_name(1)

#> [1] "FruMARCM-M002262_seg001"

# "FruMARCM-M002262_seg001"
isTRUE(fc_idid("FruMARCM-M002262_seg001")==1)

#> [1] TRUE

isTRUE(fc_neuron("FruMARCM-M002262_seg001")=="fru-M-200266")

#> [1] TRUE