Plot exemplar neurons/all cluster members after clustering by affinity

For each cluster, savefun receives the exemplar, the other neurons for the current cluster, all exemplars and the cluster_id (numeric). selected is a list with 0 or 1 named entries for each neuron. An entry can contain up to 3 fields: * keepcluster (TRUE/FALSE) * keep (character vector of neurons) * reject (character vector of neurons)

# S3 method for APResult
plot3d(x, plot = c("exemplars", "bycluster", "all"),
  suppressPlot = FALSE, savefun = NULL, clusters = NULL, soma = TRUE,
  Verbose = TRUE, selected = list(), selected_file = NULL, yaml = TRUE,
  col = NULL, ...)

Arguments

x

An APResult object from apcluster

plot

Whether to plot exemplars or all members of each cluster

suppressPlot

logical indicating that plots should be suppressed.

savefun

Optional function called after each plot is shown

clusters

Names of subset of clusters to plot

soma

Whether to plot somata (default TRUE)

Verbose

logical indicating that output should be verbose.

selected

A list selecting some of the neurons from each cluster (see Details)

selected_file

an optional path to a yaml file to load a selection from.

yaml

logical indicating that selections should be saved as yaml files, not rda files.

col

Optional argument passed to plot3dfc.

...

options passed to plot3dfc when plot=exemplars

Value

list with names of selected neurons for each cluster

See also

apcluster

Examples

## Not run: ------------------------------------
# myselection=plot3d(apres15kv2ns,'bycluster')
# myselection=LoadObjsFromRda('myselectionfile.rda')[[1]]
# mynewselection=plot3d(apres15kv2ns,'bycluster',selected=myselection)
## ---------------------------------------------