Plot FlyCircuit neurons using rgl

Plot FlyCircuit neurons using rgl

plot3dfc(id, col, db = getOption("nat.default.neuronlist"), ...)

Arguments

id
Vector of FlyCircuit gene or neuron names or idids (passed to fc_gene_name)

col
Function or vector specifying colour.

db
Object of class neuronlist or a character vector naming a neuronlist that will be passed to link{get}.

...
Additional arguments for plot3d

Value

List of rgl stack ids (see rgl::plot3d).

Details

Colour defaults to black for a single neuron or rainbow(length(id)) when more than one neuron is plotted. It can be specified as a function, a vector (length(id)) or a single recycled value.

By default when more than 200 neurons are plotted in a single call rgl redraw is suspended. This can be >10x faster.

npop3d can be used to remove the last plotted neurons.

Plotted soma positions (when soma=TRUE) is passed as an additional argument for plot3d.neuronlist are a prediction based on location in the FlyCircuit template brain mapped onto my selected template. The positions are looked up in the df attribute of the neuronlist.

See also

plot3d, pop3dfc, fc_gene_name

Examples

open3d()

#> glX 
#>   2 

library(nat)
plot3dfc("fru-M-300145", col='green', db=kcs20)
npop3d()
plot3dfc("fru-M-100014", col='red', db=kcs20, soma=TRUE)
rgl.close()